+++ rare/urls.py
@@ -17,7 +17,7 @@ Including another URLconf
 from django.contrib import admin
 from django.urls import include, path
 from rest_framework.routers import DefaultRouter
-from rareapi.views import login_user, CategoryViewSet, CommentView, TagView, UserViewSet
+from rareapi.views import CategoryViewSet, CommentView, TagView, UserViewSet
 
 
 router = DefaultRouter(trailing_slash=False)
@@ -28,6 +28,6 @@ router.register(r"tags", TagView, "tag")
 urlpatterns = [
     path("", include(router.urls)),
     path('admin/', admin.site.urls),
-    path('login', login_user),
+    path('login', UserViewSet.login_user),
     path('register', UserViewSet.as_view({'post': 'register_account'}), name='register'),
 ]
