+++ rareapi/views/__init__.py
@@ -1,5 +1,4 @@
 from .users import UserViewSet
-from .auth import login_user
 from .categories import CategoryViewSet
 from .comments import CommentView
 from .tags import TagView
