--- rareapi/views/auth.py
@@ -1,38 +1,0 @@
-from rest_framework.decorators import api_view, permission_classes
-from rest_framework.permissions import AllowAny
-from django.contrib.auth import authenticate
-from django.contrib.auth.decorators import login_required
-from rest_framework.authtoken.models import Token
-from rest_framework.response import Response
-
-
-@api_view(['POST'])
-@permission_classes([AllowAny])
-def login_user(request):
-    '''Handles the authentication of a user
-
-    Method arguments:
-      request -- The full HTTP request object
-    '''
-    username = request.data['username']
-    password = request.data['password']
-
-    # Use the built-in authenticate method to verify
-    # authenticate returns the user object or None if no user is found
-    authenticated_user = authenticate(username=username, password=password)
-
-    # If authentication was successful, respond with their token
-    if authenticated_user is not None:
-        token = Token.objects.get(user=authenticated_user)
-
-        data = {
-            'valid': True,
-            'token': token.key,
-            'staff': token.user.is_staff
-        }
-        return Response(data)
-    else:
-        # Bad login details were provided. So we can't log the user in.
-        data = { 'valid': False }
-        return Response(data)
-
